- renamed sessionProvider back to sessionManagaer
